練習中

+++ src/main/java/com/example/demo/entity/Task.java
@@ -1,6 +1,8 @@
 package com.example.demo.entity;
 
-import java.sql.Date;
+
+
+import java.util.Date;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -10,22 +12,20 @@ import javax.persistence.Table;
 
 @Entity
 @Table(name="task")
-
 public class Task {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
+	private Long id;
 	private String title;
 	private Date create_date;
 	private Date update_date;
 	private int is_Done;
 
-
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getTitle() {
